chore(monitoring): consolidate tests + drop noisy comments
- Merge judgeChange.test, diff-orchestrator.test, page-events.test
  into one meaningful-monitoring.test.ts. Cuts redundant cases
  (e.g. the derivePageWebhookEvents matrix collapses to one assertion).
- Remove restate-the-what comments throughout. Keep only WHY comments
  (truncation safety, forward-compat insert) and trim those.
