+++ apps/api/src/services/monitoring/runner.ts
@@ -611,9 +611,6 @@ async function sendNotifications(params: {
     }
   }
 
-  // Pass the full filtered page list so the judgment-gating logic in
-  // sendMonitoringEmailSummary sees every changed page. The email
-  // renderer caps the list for display on its own.
   const emailStatus = await sendMonitoringEmailSummary({
     monitor: params.monitor,
     check: params.check,
